% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curl_download2.R
\name{curl_download2}
\alias{curl_download2}
\title{curl_download2}
\usage{
curl_download2(..., .sleep)
}
\arguments{
\item{...}{path through `curl::curl_download()`}

\item{.sleep}{sleep time.}
}
\description{
curl_download2
}
