% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_construct_blastdb.R
\name{db_construct_blastdb}
\alias{db_construct_blastdb}
\title{Construct BLASTDB.}
\usage{
db_construct_blastdb(.seq_path, .save_dir = "blastdb")
}
\arguments{
\item{.seq_path}{path of fasta files.}

\item{.save_dir}{save directory.}
}
\description{
Construct BLASTDB.
}
