% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_convert_taxid2acc.R
\name{db_convert_taxid2acc}
\alias{db_convert_taxid2acc}
\title{Convert taxid to accession ids.}
\usage{
db_convert_taxid2acc(.taxid, .db_path)
}
\arguments{
\item{.taxid}{taxids.}

\item{.db_path}{database path.}
}
\description{
Convert taxid to accession ids.
}
