% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_convert_taxon2acc.R
\name{db_convert_taxon2acc}
\alias{db_convert_taxon2acc}
\title{Convert taxon name to accession ids.}
\usage{
db_convert_taxon2acc(
  .taxon,
  .target_rank,
  .db_path_tax,
  .db_path_acc,
  .check_kingdom = "Fungi",
  .lower_taxon = TRUE
)
}
\arguments{
\item{.taxon}{taxon name.}

\item{.target_rank}{taxon rank.}

\item{.db_path_tax}{taxonomy database path.}

\item{.db_path_acc}{accession database path.}

\item{.check_kingdom}{kingdom of taxon.}

\item{.lower_taxon}{get taxid in lower taxon.}
}
\description{
Convert taxon name to accession ids.
}
