% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_divide_acc.R
\name{db_divide_acc}
\alias{db_divide_acc}
\title{Classify accession ids.}
\usage{
db_divide_acc(.acc, .type = "Direct submissions")
}
\arguments{
\item{...}{path through `curl::curl_download()`}

\item{.sleep}{sleep time.}
}
\description{
Classify accession ids.
}
