% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_extract_ITS.R
\name{db_extract_ITS}
\alias{db_extract_ITS}
\title{Extract ITS region by using ITSx}
\usage{
db_extract_ITS(
  .raw_fasta,
  .marker_region,
  .save_dir,
  .db_name,
  .type = c("full", "marker"),
  .nthread = 10
)
}
\arguments{
\item{.raw_fasta}{path of fasta files.}

\item{.marker_region}{marker region (ITS1, ITS2, 5.8S, ITS, SSU, LSU, all).}

\item{.save_dir}{save directory.}

\item{.db_name}{names for output fasta files.}

\item{.type}{return marker region only or full length sequences.}

\item{.nthread}{number of thread}
}
\description{
Extract ITS region by using ITSx
}
