% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_get_meta.R
\name{db_get_meta}
\alias{db_get_meta}
\title{Get metadata as XML format}
\usage{
db_get_meta(.acc, .dest_dir, .n_max, .sleep, .try_max = NULL)
}
\arguments{
\item{.acc}{accession ids.}

\item{.dest_dir}{save path.}

\item{.n_max}{maximum number of chunks.}

\item{.sleep}{sleep time.}

\item{.try_max}{number of trials.}
}
\description{
Get metadata as XML format
}
