% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_parse_xml.R
\name{db_parse_gbxml}
\alias{db_parse_gbxml}
\title{Parse genbank xml files.}
\usage{
db_parse_gbxml(.target_dir, .gbseq, .workers)
}
\arguments{
\item{.target_dir}{target directory that contains xml fiels.}

\item{.gbseq}{parse option.}

\item{.workers}{multisession. see `?future::multisession`.}
}
\description{
Parse genbank xml files.
}
