% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_remove_chimera.R
\name{db_remove_chimera}
\alias{db_remove_chimera}
\title{Remove chimeric sequences.}
\usage{
db_remove_chimera(
  .save_dir = "phylogeny",
  .marker_region = c("full", "ITS1", "ITS2", "LSU", "SSU"),
  .ref_db_path
)
}
\arguments{
\item{.save_dir}{directory for save.}

\item{.marker_region}{maker}

\item{.ref_db_path}{UCHIME reference db path}
}
\description{
Remove chimeric sequences.
}
