% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_remove_outlier.R
\name{db_remove_outlier}
\alias{db_remove_outlier}
\title{Remove outliner.}
\usage{
db_remove_outlier(.save_dir = "phylogeny", .dont_use_python = T, .nthread = 10)
}
\arguments{
\item{.save_dir}{save directory.}

\item{.dont_use_python}{use python or not.}

\item{.nthread}{number of threads}
}
\description{
Remove outliner.
}
