% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_remove_redandancy.R
\name{db_remove_redundany}
\alias{db_remove_redundany}
\title{Clustering.}
\usage{
db_remove_redundany(
  .work_dir = "phylogeny",
  .fas,
  .centroid = "clustered.fas",
  .consensus = "consensus.fas",
  .save_dir = "cluster_dir",
  .id = 0.98
)
}
\arguments{
\item{.work_dir}{working directory.}

\item{.marker_region}{maker}

\item{.ref_db_path}{UCHIME reference db path}
}
\description{
Clustering.
}
