% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_select_acc.R
\name{db_select_acc}
\alias{db_select_acc}
\title{Select accession ids.}
\usage{
db_select_acc(.acc, .type = "Direct submissions")
}
\arguments{
\item{...}{path through `curl::curl_download()`}

\item{.sleep}{sleep time.}
}
\description{
Select accession ids.
}
