% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_set_acc2taxid.R
\name{db_set_acc2taxid}
\alias{db_set_acc2taxid}
\title{Donwload accession2taxid}
\usage{
db_set_acc2taxid(.db_dir, .override = TRUE, .save_to_parquet = TRUE)
}
\arguments{
\item{.db_dir}{directory.}

\item{.override}{override old one.}

\item{.save_to_parquet}{convert to parquet format.}
}
\description{
Donwload accession2taxid
}
