% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_set_acc2taxid_.R
\name{db_set_acc2taxid_}
\alias{db_set_acc2taxid_}
\title{Donwload accession2taxid_}
\usage{
db_set_acc2taxid_(.db_dir, .override = TRUE, .save_to_parquet = TRUE)
}
\arguments{
\item{.db_dir}{directory.}

\item{.override}{override old one.}

\item{.save_to_parquet}{convert to parquet format.}
}
\description{
Donwload accession2taxid_
}
