% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_set_taxdmp.R
\name{db_set_taxdmp}
\alias{db_set_taxdmp}
\title{Donwload taxdmp.}
\usage{
db_set_taxdmp(.db_dir = "db", .override = TRUE, .save_to_parquet = TRUE)
}
\arguments{
\item{.db_dir}{directory.}

\item{.override}{override old one.}

\item{.save_to_parquet}{convert to parquet format.}
}
\description{
Donwload taxdmp.
}
