% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promap_dfr.R
\name{promap}
\alias{promap}
\title{map() with progress bar.}
\usage{
promap(.x, .f, ...)
}
\arguments{
\item{.x}{variable 1.}

\item{.f}{function.}

\item{...}{augments.}

\item{.y}{variable 2.}
}
\description{
heavily inspired with https://github.com/tidyverse/purrr/issues/149#issuecomment-1167456412
}
