% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promap2_chr.R
\name{promap2_chr}
\alias{promap2_chr}
\title{map2_chr() with progress bar.}
\usage{
promap2_chr(.x, .y, .f, ...)
}
\arguments{
\item{.x}{variable 1.}

\item{.y}{variable 2.}

\item{.f}{function.}

\item{...}{augments.}
}
\description{
heavily inspired with https://github.com/tidyverse/purrr/issues/149#issuecomment-1167456412
}
