% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furrr2.R
\name{supmap}
\alias{supmap}
\title{Suppress message of furrr::future_map()}
\usage{
supmap(.x, .f, ...)
}
\arguments{
\item{.x}{list.}

\item{.f}{function.}

\item{...}{path through `future_map()`}
}
\description{
Suppress message of furrr::future_map()
}
