% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_parse_xml.R
\name{db_parse_gbxml}
\alias{db_parse_gbxml}
\title{Parse genbank xml files.}
\usage{
db_parse_gbxml(.target_dir)
}
\arguments{
\item{.target_dir}{target directory that contains xml fiels..}
}
\description{
Parse genbank xml files.
}
